[WOF-US05] Player.java moved to correct package
